# Conflicts:
#	frontend/src/features/arrivals/components/ArrivalsDataList.tsx
#	frontend/src/features/stocks/containers/StockPage.tsx

+++ frontend/src/features/orders/hooks/useOrderPage.ts
@@ -15,7 +15,6 @@ const UseOrderPage = () => {
   const orders = useAppSelector(selectAllOrdersWithClient)
   const loading = useAppSelector(selectLoadingFetchOrder)
   const [open, setOpen] = useState(false)
-  const [counterpartyToDelete, setCounterpartyToDelete] = useState<OrderWithClient | null>(null)
 
   useEffect(() => {
     dispatch(fetchOrdersWithClient())
@@ -23,11 +22,13 @@ const UseOrderPage = () => {
 
   const handleDelete = async (id: string) => {
     try {
-      await dispatch(deleteOrder(id))
-      dispatch(fetchOrdersWithClient())
-      toast.success('Заказ успешно удалён!')
+      if (confirm('Вы уверены, что хотите удалить этот заказ?')) {
+        await dispatch(deleteOrder(id))
+        dispatch(fetchOrdersWithClient())
+      } else {
+        toast.info('Вы отменили удаление заказа')
+      }
     } catch (e) {
-      toast.error('Ошибка при удалении заказа.')
       console.error(e)
     }
   }
@@ -37,7 +38,7 @@ const UseOrderPage = () => {
   const handleClose = () => {
     setOpen(false)
     dispatch(clearPopulateOrder())
-    dispatch(clearErrorOrder())
+    dispatch( clearErrorOrder ())
   }
 
   const handleOpenEdit = async (order: OrderWithClient) => {
@@ -45,13 +46,6 @@ const UseOrderPage = () => {
     setOpen(true)
   }
 
-  const handleConfirmDelete = async () => {
-    if (counterpartyToDelete) {
-      await dispatch(deleteOrder(counterpartyToDelete._id))
-      dispatch(fetchOrdersWithClient())
-      handleClose()
-    }
-  }
 
   return {
     orders,
@@ -61,8 +55,6 @@ const UseOrderPage = () => {
     handleClose,
     handleDelete,
     handleOpenEdit,
-    setCounterpartyToDelete,
-    handleConfirmDelete,
   }
 }
